Added increment of callstack offset
Also some code reorganization.
